Server side render Home page to support sprig
